requirements.txt
google-cloud-aiplatform
python-dotenv
google-genai
streamlit